ISSUE #366 moved testClient
